Created the concept of features which have actions
The big idea, here, is that I wasn't happy with the highly specific feature architecture I was building. I wanted something more flexible, more understandable, and more tied to interactive fiction. Thus, the feature/action syntax displayed here. This is a complete rearchitecture, that is both simpler and more flexible.