Refined the behavior of the features.
Mostly, I'm cleaning up the DSL syntax, adding some methods for convenient access, renaming methods for better clarity, and basically trying to make this fucking work.

It's not. Not included in this commit is a set of failing tests. I wanted to get something up that could be reviewed, but it's still a loooong way from primetime.