Proper generation of handle_call methods.
Okay, so I was having an issue with actions- I wanted to have each action generate a handle_call method, but then I ran into an issue where the macro failed to expand properly, and thus I couldn't successfully pattern match in the function signature.

I've fixed that by switching to using a module attribute to guide that compilation. I'm not 100% happy with how this ends up looking, but it does the job, and it gives me the scene description syntax I want.