Initializing Feature State
As it says on the tin, we can now initialize a feature's state using a `setup` macro, which in turn depends on the `set_state` macro… which works fine in this case. I'm having issues with `get_state` in some other cases, so the implementation of these macros might change dramatically between now and then.