Converted Actor to a parameter.
Initially, I was passing actor as the sender, under the assumption that the Actor would be a front-end to any messages being passed to a Scene. This was a silly assumption. For example: in any test I write, I'm not testing a goddamn scene and actor combination, because I'm trying to write a unit test. Now my code is less stupid.